FDKv2 ubsan patches
Bug: 80053205
Test: see bug for repro with FB "wow"
   atest DecoderTestAacDrc

Fix signed integer overflows in CLpc_SynthesisLattice()


Fix signed integer overflows in imlt


Fix overflow in addLowbandEnergies()


Concealment fix for audio frames containing acelp components


Fix out-of-bounds access in PS concealment


Fix potential memory leak in setup of qmf domain


Reject channel config 13


Fix unsigned integer overflow in configExtension()


Fix unsigned integer overflow in CAacDecoder_DecodeFrame()


Fix unsigned integer overflow in aacDecoder_UpdateBitStreamCounters()


Fix unsigned integer overflow in transportDec_readStream()

+++ libMpegTPDec/src/tpdec_lib.cpp
@@ -1337,9 +1337,9 @@ static TRANSPORTDEC_ERROR transportDec_readStream(HANDLE_TRANSPORTDEC hTp,
   INT bitDistance, bfDelta;
 
   /* Obtain distance to next synch word */
-  bitDistance = FDKgetValidBits(hBs);
+  bitDistance = (INT)FDKgetValidBits(hBs);
   error = synchronization(hTp, &headerBits);
-  bitDistance -= FDKgetValidBits(hBs);
+  bitDistance -= (INT)FDKgetValidBits(hBs);
 
   FDK_ASSERT(bitDistance >= 0);
 
@@ -1380,7 +1380,7 @@ static TRANSPORTDEC_ERROR transportDec_readStream(HANDLE_TRANSPORTDEC hTp,
         int num, denom;
 
         /* Obtain estimate of number of lost frames */
-        num = hTp->asc[0].m_samplingFrequency * (bfDelta + bitDistance) +
+        num = (INT)hTp->asc[0].m_samplingFrequency * (bfDelta + bitDistance) +
               hTp->remainder;
         denom = hTp->avgBitRate * hTp->asc[0].m_samplesPerFrame;
         if (num > 0) {
